feat: update dtos

+++ src/main/java/com/pique/pique/domain/dtos/transaction/TransactionDTO.java
@@ -1,8 +1,10 @@
 package com.pique.pique.domain.dtos.transaction;
 
+import java.math.BigDecimal;
+
 public record TransactionDTO(
         Long senderId,
         Long receiverId,
-        Double value
+        BigDecimal value
 ) {
 }
