+++ src/main/java/com/pique/pique/domain/dtos/transaction/TransactionDTOResponse.java
@@ -3,12 +3,13 @@ package com.pique.pique.domain.dtos.transaction;
 import com.pique.pique.domain.dtos.user.UserDTOResponse;
 import com.pique.pique.domain.entities.TransactionEntity;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 public record TransactionDTOResponse(
         UserDTOResponse senderId,
         UserDTOResponse receiverId,
-        Double value,
+        BigDecimal value,
         LocalDateTime timestamp
 ) {
     public static TransactionDTOResponse fromTransactionEntity(TransactionEntity transactionEntity) {
